Spins up threads per prob dist.
Should be caluculating probabilities correctly for a singe ply. Haven't
gotten to merging or writing back into the tree yet. That will come
tomorrow. Also all of the stuff that is important for mathing in the
threads is now only in C land. So no ruby accesses happen inside the
threads.